Added connection to firebase but
