--- src/pages/Graphs/styles.css
@@ -1,21 +1,0 @@
-.App {
-    font-family: sans-serif;
-    text-align: center;
-}
-
-.tile {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    /*color: white;*/
-    /*background-color: #4b6cb7;*/
-    box-shadow: rgba(0, 0, 0, 0.05) 3px 3px 15px;
-    transition: all 300ms ease-in-out;
-}
-
-.dragging {
-    transform: scale(1.06);
-    box-shadow: rgba(0, 0, 0, 0.25) 0px 54px 55px,
-    rgba(0, 0, 0, 0.12) 0 -12px 30px, rgba(0, 0, 0, 0.12) 0 4px 6px,
-    rgba(0, 0, 0, 0.17) 0 12px 13px, rgba(0, 0, 0, 0.09) 0 -3px 5px;
-}
