+++ src/pages/dashboard/test.tsx
@@ -1,0 +1,7 @@
+export default function test() {
+    return (
+        <>
+            Hi
+        </>
+    );
+}
