+++ src/router.tsx
@@ -5,7 +5,7 @@ import IndexPage from "./pages/index/IndexPage.tsx";
 import {SampleSidebar} from "./pages/samples/sidebar/SampleSidebar.tsx";
 import {SampleSidebarHeader} from "./pages/samples/sidebar_header/SampleSidebarHeader.tsx";
 import {SampleModal} from "./pages/samples/modal/SampleModal.tsx";
-import GraphDashboard from "./pages/Graphs/Graphs.tsx";
+import GraphDashboard from "./pages/dashboard/DashboardPage.tsx";
 import {TestFirestorePage} from "./pages/test firestore/TestFirestore.tsx";
 
 // ? Routing - see https://reactrouter.com/en/main
